API Reference

This section provides the full API documentation for Moose.jl.

Types

Moose.Basis — Type

Basis()

Description

This struct loads the basis vectors matrix H and pre-computes several matricial products and stores them in a julia Struct. This storage allows to save computational ressources and speed the code. No arguments are needed to initialize this struct. By default it loads a rank 10 NMF basis vectors obtained using a sequentail nearly-NMF on ~7000 galaxy spectra

Arguments

None

Details

Moose.Γgrid — Type

Γgrid(; λmin = 4500f0, λmax = 9350f0, δλ = 1.25f0,
        ζmin = 0f0,    ζmax = 7f0,    δζ = 0.0015f0)

Description

Creates a rest-frame log-wavelength grid regularly spaced. And provides the test redshifts vector. All needed arguments to initialize this struct have default values, so that, one can get the grid without specifying any argument. Note that the resolution δΓ of the grid is not an argument, if one wants to play with this parameter you have to change the code. δΓ is calculated as the mean resolution of the rest-frame wavelength grid not regularly spaced: log10(min observed wavelength/ (1+maximum redshift) : instrument resolution / (1+maximum redshift)): maximum observed wavelength.

Arguments

λmin ::Float32 = 4500 : MUSE minimal observed wavelength in Å

λmax ::Float32 = 9350 : MUSE maximal observed wavelength in Å

δλ ::Float32 = 1.25 : MUSE spectral resolution in Å

ζmin ::Float32 = 0 : minimum test redshift

ζmax ::Float32 = 7 : maximum test redshift

δζ ::Float32 = 0.0015 : step of test redshift

Details

This structure fills additional fields:

-δΓ :: Float32 : rest-frame spectral resolutionm -Γ :: Vector{Float32} : rest-frame grid log-wavelengths vector -ζ :: Vector{Float32} : vector of test redshifts

Examples

# to get the rest-frame log-wavelength grid just run
Γ = Γgrid().Γ

Metrics

Moose.GF — Function

GF(z::Vector{Float32},ẑ::Vector{Float32}; tMUSE::Float32 = 0.1f0)

Description

This function calculates the Good Fraction (GF), given vectors of true and predicted redshift z and ẑ. GF is defined as,

\[\begin{aligned} && GF &= \frac{\rm good}{N} \\ \end{aligned}\]

$N_{\rm good}$ and $N$ are respectively the number of good predictions and the total number of predictions. A prediction is good when it satistifies $\Delta z = |z - ẑ| < t$, where $t$ is a threshold that can be set by setting the keyword argument tMUSE.

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

tMUSE: redshift error tolerance for MUSE

Details

Given the vectors of true redshifts z and the predicted redshifts ẑ, this function calculates the good fraction GF, defined as the ratio between the number of good predictions and the total number of predictions. A good prediction is a one that has an error Δz = |z - ẑ| less than a threshold t (tMUSE) in the code, the threshold is set by the scientific goal.

Author(s)

B. masten

Moose.MAD — Function

MAD(z::Vector{Float32}, ẑ::Vector{Float32})

Description

This function calculates the median absolute deviation (MAD), given the vectors of true and predicted resdshifts (z & ẑ respectively). MAD is calculated as follow: MAD = median(|Δz_i - median(Δz)|)

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

Author(s)

B. Masten

Moose.MAE — Function

MAE(z::Vector{Float32}, ẑ::Vector{Float32})

Description

This function calculates the mean absolute error (MAE) score, given the vectors of true and predicted resdshifts (z & ẑ respectively). MAE is calculated as follow: MAE = (Σ_i i=1..N |z - ẑ|) / N

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

Author(s)

B. Masten

Moose.Δχ² — Function

Δχ²(χ²::Vector{Float32})

Description

This function calculates the Δχ² significance score, given a chi-square curve Δχ² is calculated as follow:

\[egin{aligned} && \Delta \chi^2 & = 1 - rac{\chi^2_{ m min}}{ m{Q1}(\chi^2)} \ nd{aligned}\]

Arguments

χ²: chi-square curve vector

Returns

Δχ²: redshift significance score

Author(s)

B. Masten

Moose.R — Function

R(χ2::Vector{Float32})

Description

This function calculates the R robustness score, given a chi-square curve R is calculated as follow:

\[egin{aligned} && R & = 1 - rac{\chi^2_{ m min}}{ m{Q1}(\chi^2)} \ nd{aligned}\]

Arguments

χ2: chi-square curve vector

Returns

R: redshift robustness score

Author(s)

B. Masten

Moose NMF

Moose.nNMF — Type

nNMF(X::Matrix{Float32}; W=nothing, H=nothing, V=nothing, M=nothing, k::Int=5, maxiters::Int=5000, tol::Float32=1f-8, verbose::Bool=true)

Description

nNMF is a type, holds X, W, H and V matrices and other algorithmic settings for the nearly-NMF algorithm solver.

Arguments

X ::Matrix{Float32} : Data matrix (n by l)

W ::Matrix{Float32} : Weights matrix (n by k), if nothing W is randomly initialized. Defaults to nothing

H ::Matrix{Float32} : Basis vectors matrix (k by l), if nothing H is randomly initialized. Defaults to nothing

V ::Matrix{Float32} : Inverse variance matrix (n by l), if nothing V is randomly initialized with ones. Defaults to nothing

M ::Matrix{Float32} : Mask matrix (n by l), if X[j,j] is missing then M[i,j] = true. if nothing M is constructed from X. Defaults to nothing

k ::Int : NMF decomposition rank (or number of basis vectors in H). Defaults to 10

maxiters ::Int : Maximum number iterations for the solver. Defaults to 5000

tol ::Float32 : Stopping criterion, if improvement < tol and the maximum iterations is not reached, solver stops. Defaults to 1f-8

verbose ::Bool : If true, algorithmic settings and the proportion of negative/missing values in X are printed. Defaults to true

Details

Element-type of all matrices (X,W,H and V) should be Float32, this is to reduce memory storage and speed calculations.

Author(s)

B. masten

Moose.nearly! — Function

nearly!(nmf::nNMF)

Description

This is a mutating function! runs the nearly NMF solver on the data matrix nmf.X, and updates nmf.W & nmf.H in-place. nearly NMF tries to solve:

\[\begin{aligned} & \quad \quad \quad \min_{W\in \mathbb{R}^{n \times k}, H\in \mathbb{R}^{k \times l}} &&||(X+Y) - (WH + Y)||_{V}^2\\ & \quad \quad \quad \text{\quad \quad s.t.} && W \geq 0 \;\text{and} \; H \geq 0 \end{aligned}\]

Y holds the minimum shift to make each entry in X non-negative. To minimize this objective, the nearly NMF solver iteratively updates $W$ and $H$ matrices with the following rules:

\[\begin{array}{l} H = H \odot \dfrac{\left[ W^{T} (V \odot X) \right]^{+}}{ W^{T}(V \odot (WH)) + \left[W^{T}(V \odot X) \right]^{-}}\\ W = W \odot \dfrac{\left[ (V \odot X)H^{T} \right]^{+}}{ (V \odot (WH))H^{T} + \left[(V \odot X)H^{T} \right]^{-}} \\ \end{array}\]

Where $\odot$ is the element-wise product. Operation $[\;]^{+}$ applied on a matrix results in a matrix with the same shape, in which all negative values are zeroed. $[\;]^{-}$ on the other hand, zeroes the positive values and takes the absolute value of negative values.

Arguments

nmf::nNMF: nNMF Struct type

Returns

χ2 : reconstruction error (frobenius norm) at last iteration

iter : exit/last iteration

Reference

Reference: Dylan Green and Stephen Bailey 2024

Moose methods

Moose.fnnls — Function

fnnls

Description

Implementation of the Fast Non-Negative Least Squares algortithm (FNNLS) adopted from the NonNegLeastSquares.jl package, (see link[https://doi.org/10.1002/(SICI)1099-128X(199709/10)11:5<393::AID-CEM483>3.0.CO;2-L]).

Problem under consideration: Ax = b, subject to x >= 0

FNNLS algorithm solves this problem in the least squares sense, || Ax - b||^2, subject to x >= 0

Arguments

AtA ::Matrix{Float32} : Gram matrix (A transpose A)

Atb ::Matrix{Float32} : Gram matrix (A transpose b)

tol ::Float32 = 1f-8 : tolerance for nonnegativity constraints

max_iter ::Int= 30*size(AtA,2) : Maximum number of iterations)

Returns

x̂ ::Vector{Float32} : x̂ that solves A*x = b in the least-squares sense, subject to `x >=0

References

Bro R, De Jong S. A fast non-negativitity-constrained least squares algorithm. Journal of Chemometrics. 11, 393–401 (1997)

Credit

Tim Holy, (NonNegLeastSquares.jl)

Moose.interpolate — Function

interpolate( basis::Basis, fλ::Vector{Float32}, σλ::Vector{Float32}, λ::Vector{Float32}; extrpfλ = 0.0f0, extrpσλ = 1f6)

Description

This function takes three vectors: flux (fλ), standard deviations (σλ) and observed wavelengths (λ), it interpolates the flux and std vectors to the rest wavelength grid assuming redshift 0. (shifting is captured in gram matrices stored in basis Struct)

Arguments

basis ::Basis : Basis Struct

fλ ::Vector{Float32} : Observed flux densities

σλ ::Vector{Float32} : Flux associated standard deviations

λ ::Vector{Float32} : Observed wavelengths

extrpfλ ::Float32 = 0f0 : Extrapolation value for fluxes

extrpσλ ::Float32 = 1f6 : Extrapolation value for standard deviations

None

returns

intrpfλ ::Vector{Float32} : Interpolated flux densities

intrpσλ ::Vector{Float32} : Interpolated standard deviations

Details

Moose.threaded_fnnls — Function

threaded_fnnls(basis::Basis, fλ::Vector{Float32}, σλ::Vector{Float32})

Description

This function executes a multi-threaded calculation of the chi-square curve. For each redshift (each gram matrices Hi & HHti in basis Struct), it reconstructs an input vector fλ with basis vectors using Fast Non-Negative Least Squares (FNNLS), then quantifies the reconstruction error using a χ2 metric

#put definition of chi2 here grid = Γgrid()

Arguments

basis ::Basis : Basis Struct

fλ ::Vector{Float32} : Input vector (flux densities)

σλ ::Vector{Float32} : Error vector (standard deviations)

Returns

χ2 ::Vector{Float32} : χ2 curve, χ2 metric at each test redshift

Example

# 1. Initialize the Basis Struct
basis = Basis()

# 2. Interpolate flux densities fλ and σλ to rest frame grid
intrpfλ, intrpσλ = interpolate(b, fλ .* λ, σλ .* λ, λ)

# 3. Run threaded_nnls to get the chi2 curve
χ2 = threaded_fnnls(b, intrpfλ, intrpσλ)

Tips

To set the number of threads in julia, add this line to your linux/macos bash

export JULIA_NUM_THREADS=4

Moose.χcube — Function

χcube(path::String, save_path::Union{String, Nothing}; kwargs...)

Description

Reads the data from a datacube in path, and runs a threaded_fnnls on each spaxel. It then saves the resulting chi2 Array into an ".h5" file.

Arguments

path ::String : Path to the datacube ".fits" file

save_path ::String : Path where to save the h5file output

Keyword Arguments

DataExtName ::String : Name of the data extension

StatExtName ::String : Name of the stat extension

Note

Keyword arguments are passed directly to leggere_cube. If the Data extension is named "DATA", there is no need to pass this keyword argument. Similarly, if the stat extension is named "STAT", there is no need to pass this keyword

Output

An ".h5" file containing the chi2 Array, located in save_path, the key to the h5 dataset is given by

h5key = splitext(basename(path))[1]

Returns

nothing

Author(s)

B.Masten

See also χfits

Moose.χfits — Function

χfits(wgrid::Γgrid, basis:Basis, data; output_path::String = nothing)

Description

Runs a threaded FNNLS on a set of spectra loaded into a data Struct apriori and stored in data. Resulting chi2 curves are saved into an h5 file, whose path and name are specified by setting the keyword argument 'output_path'

Arguments

wgrid ::Γgrid : Γgrid Struct

basis ::Basis : Basis Struct

data ::Struct : Named Tuple with four fields: flux, sdev, wave, ids

output_path ::String : Path where the chi2 file will be saved

Returns

nothing

Methods

χfits(data; output_path::String = nothing): wgrid and basis are instantiated within the method.

χfits(fits_path::String, DataExtName::Union{String, Int}, StatExtName::Union{String, Int}; output_path::String = nothing): wgrid and basis are instantiated within the method, the data is loaded within the function using the provided fits_path and the extensions: DataExtName, StatExtName.

χfits(wgrid::Γgrid, basis::Basis, fits_path::String, DataExtName::Union{String, Int}, StatExtName::Union{String, Int}; output_path::String = nothing): the data is loaded within the function using the provided fits_path and the extensions: DataExtName, StatExtName.

Example

# 1. Initialize Basis Struct
wgrid = Γgrid()
basis = Basis()

# 2. read data from fits files
data = leggere_fits("../data/spectra_sample")

#3. Specify output file path and run χloop
my_chi2file = "../output/chi2_files/chi2_moose.h5"
χfits(wgrid, basis, data; output_path = my_chi2file)

Utils

Moose.sanitize — Function

sanitize(z,ẑ)

Description

Help function, this function checks for any NaN or Inf in the input vectors of true and predicted redshifts (z & ẑ respectively), if any it masks them and raises a warning to the user. It also ensures that both vectors have the same size.

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

Author(s)

B. Masten

Moose.leggere_fits — Function

leggere_fits(path::String; DataExtName::String= "DATA", StatExtName::String ="STAT")

Description

=========== This function reads fluxes, standard deviations, wavelengths and ids from a group of fits file in 'path', and outputs a Struct with four fields (flux, sdev, awave, ids). The name of the data and stat extension can be specied by setting keywords 'DataExtName' and 'StatExtName'

Arguments

path ::String : Path to the fits files folder

DataExtName ::String = "DATA" : Name of the data extension

StatExtName ::String = "STAT" : Name of the stat extension

returns

data ::Struct : Struct with four fields (flux, sdev, awave, ids)

warning `DataExtName` and `StatExtName` have to be consistent across fits files`

Moose.leggere_chifile — Function

leggere_chifile(path::String)

Description

This function reads χ² curves, best decomposition parameters ω , predicted redshift ẑ, significance score Δχ², and robustness score R, stored in a .h5 file and to which the argument path points at.

Arguments

path ::String : Path to the h5 file

returns

data ::NamedTuple : NamedTuple with 6 fields (ids, chi2, omega, zhat, dchi2, R)

Moose.leggere_cube — Function

leggere_cube(path::String; DataExtName::String ="DATA", StatExtName::String = "STAT" )

Description

This function reads the flux densities, standard deviations, wavelength array and attributes and id to each spaxel in the datacube fits file in path. It outputs a NamedTuple with four fields (flux, sdev, awave, ids). The name of the data and stat extension in the fits file can be specified by setting keywords arguments DataExtName and StatExtName

Arguments

path ::String : Path to the fits files folder

DataExtName ::String = "DATA" : Name of the data extension

StatExtName ::String = "STAT" : Name of the stat extension

Returns

NamedTuple with four fields (flux, sdev, awave, ids)