API Reference

This section provides the full API documentation for Moose.jl.

Structs

Moose.Γgrid — Type

Γgrid(; λmin = 4500f0, λmax = 9350f0, δλ = 1.25f0,
        ζmin = 0f0,    ζmax = 7f0,    δζ = 0.001f0)

Description

This function creates a rest-frame log-wavelength grid with a uniform spacing, and provides the test redshifts vector. The function has only kerword arguments, and all of them have default values.

Arguments

λmin ::Float32 = 4600f0 : MUSE minimal observed wavelength in Å

λmax ::Float32 = 9350f0 : MUSE maximal observed wavelength in Å

δλ ::Float32 = 1.25f0 : MUSE spectral sampling in Å

ζmin ::Float32 = 0f0 : minimal test redshift

ζmax ::Float32 = 6.7f0 : maximal test redshift

δζ ::Float32 = 0.001f0 : test redshifts step

Details

This structure fills additional fields:

δΓ :: Float32 : the rest-frame log-wavelength grid uniform spacing

Γ :: Vector{Float32} : the rest-frame grid log-wavelengths vector

ζ :: Vector{Float32} : the vector of test redshifts

Examples

    # to get the rest-frame log-wavelength run
    Γ = Γgrid().Γ
    # to get the vector of test redshifts
    ζo publish  = Γgrid().ζ

Note

The grid resolution δΓ is not an input argument. To modify this parameter, the code itself must be changed. δΓ is computed as:

\[\delta \Gamma = \mathrm{mean} \left( \log\!\left(\frac{\lambda_{\mathrm{obs,min}}}{1 + z_{\mathrm{max}}}\right) : \frac{\delta\lambda}{1 + z_{\mathrm{max}}} : \lambda_{\mathrm{obs,max}} \right)\]

Moose.Basis — Type

Basis(wgrid::Γgrid; rank::Int = 10)

Description

This function instantiates a Basis Struct. It loads the basis vectors matrix H, stores slices of H corresponding to test redshifts, and pre-computes several matricial products. By default it loads a rank-10 NMF basis vectors obtained using a sequential nearly-NMF carried on ~ 7000 MUSE galaxy spectra.

Arguments

-wgrid :: Γgrid : rest-frame grid Struct

Optional arguments

-rank :: Int : rank of the basis you want to use, default to 10. Rank ∈ [6, 14]

Fields

H :: Matrix{Float32}: Full basis vectors matrix

Hᵀ :: Matrix{Float32}: Transposed basis vectors matrix

Hᵢ :: Vector{Matrix{Float32}}: Slices of H corresponding to test redshifts in wgrid.ζ

HHᵀᵢ :: Vector{Matrix{Float32}}: Hᵢ * Hᵢᵀ matricial products

Λ :: Matrix{Float32}: Rest wavelengths array corresponding to test redshifts in wgrid.ζ

k :: Int: Rank of H

l :: Int: Spectral size of basis vectors

n :: Int: Number of test redshifts

Metrics

Moose.GF — Function

GF(z::Vector{Float32},ẑ::Vector{Float32}; tMUSE::Float32 = 0.1f0)

Description

This function calculates the Good Fraction (GF), given vectors of true and predicted redshift z and ẑ. GF is defined as,

\[\begin{aligned} && GF &= \frac{\rm good}{N} \\ \end{aligned}\]

$N_{\rm good}$ and $N$ are respectively the number of good predictions and the total number of predictions. A prediction is good when it satistifies $\Delta z = |z - ẑ| < t$, where $t$ is a threshold that can be set by setting the keyword argument tMUSE.

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

tMUSE: redshift error tolerance for MUSE

Details

Given the vectors of true redshifts z and the predicted redshifts ẑ, this function calculates the good fraction GF, defined as the ratio between the number of good predictions and the total number of predictions. A good prediction is a one that has an error Δz = |z - ẑ| less than a threshold t (tMUSE) in the code, the threshold is set by the scientific goal.

Author(s)

B. masten

Moose.MAD — Function

MAD(z::Vector{Float32}, ẑ::Vector{Float32})

Description

This function calculates the median absolute deviation (MAD), given the vectors of true and predicted resdshifts (z & ẑ respectively). MAD is calculated as follow: MAD = median(|Δz_i - median(Δz)|)

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

Author(s)

B. Masten

Moose.MAE — Function

MAE(z::Vector{Float32}, ẑ::Vector{Float32})

Description

This function calculates the mean absolute error (MAE) score, given the vectors of true and predicted resdshifts (z & ẑ respectively). MAE is calculated as follow: MAE = (Σ_i i=1..N |z - ẑ|) / N

Arguments

z: vector of true redshifts

ẑ: vector of predicted redshifts

Author(s)

B. Masten

Moose.Δχ² — Function

Δχ²(χ²::Vector{Float32})

Description

This function calculates the Δχ² significance score, given a chi-square curve Δχ² is calculated as follow:

\[egin{aligned} && \Delta \chi^2 & = 1 - rac{\chi^2_{ m min}}{ m{Q1}(\chi^2)} \ nd{aligned}\]

Arguments

χ²: chi-square curve vector

Returns

Δχ²: redshift significance score

Author(s)

B. Masten

Moose.R — Function

R(χ2::Vector{Float32})

Description

This function calculates the R robustness score, given a chi-square curve R is calculated as follow:

\[egin{aligned} && R & = 1 - rac{\chi^2_{ m min}}{ m{Q1}(\chi^2)} \ nd{aligned}\]

Arguments

χ2: chi-square curve vector

Returns

R: redshift robustness score

Author(s)

B. Masten

Moose NMF

Moose.nNMF — Type

nNMF(X::Matrix{Float32}; W=nothing, H=nothing, V=nothing, M=nothing, k::Int=5, maxiters::Int=5000, tol::Float32=1f-8, verbose::Bool=true)

Description

nNMF is a type, holds X, W, H and V matrices and other algorithmic settings for the nearly-NMF algorithm solver.

Arguments

X ::Matrix{Float32} : Data matrix (n by l)

W ::Matrix{Float32} : Weights matrix (n by k), if nothing W is randomly initialized. Defaults to nothing

H ::Matrix{Float32} : Basis vectors matrix (k by l), if nothing H is randomly initialized. Defaults to nothing

V ::Matrix{Float32} : Inverse variance matrix (n by l), if nothing V is randomly initialized with ones. Defaults to nothing

M ::Matrix{Float32} : Mask matrix (n by l), if X[j,j] is missing then M[i,j] = true. if nothing M is constructed from X. Defaults to nothing

k ::Int : NMF decomposition rank (or number of basis vectors in H). Defaults to 10

maxiters ::Int : Maximum number iterations for the solver. Defaults to 5000

tol ::Float32 : Stopping criterion, if improvement < tol and the maximum iterations is not reached, solver stops. Defaults to 1f-8

verbose ::Bool : If true, algorithmic settings and the proportion of negative/missing values in X are printed. Defaults to true

Details

Element-type of all matrices (X,W,H and V) should be Float32, this is to reduce memory storage and speed calculations.

Author(s)

B. masten

Moose.nearly! — Function

nearly!(nmf::nNMF)

Description

This is a mutating function! runs the nearly NMF solver on the data matrix nmf.X, and updates nmf.W & nmf.H in-place. nearly NMF tries to solve:

\[\begin{aligned} & \quad \quad \quad \min_{W\in \mathbb{R}^{n \times k}, H\in \mathbb{R}^{k \times l}} &&||(X+Y) - (WH + Y)||_{V}^2\\ & \quad \quad \quad \text{\quad \quad s.t.} && W \geq 0 \;\text{and} \; H \geq 0 \end{aligned}\]

Y holds the minimum shift to make each entry in X non-negative. To minimize this objective, the nearly NMF solver iteratively updates $W$ and $H$ matrices with the following rules:

\[\begin{array}{l} H = H \odot \dfrac{\left[ W^{T} (V \odot X) \right]^{+}}{ W^{T}(V \odot (WH)) + \left[W^{T}(V \odot X) \right]^{-}}\\ W = W \odot \dfrac{\left[ (V \odot X)H^{T} \right]^{+}}{ (V \odot (WH))H^{T} + \left[(V \odot X)H^{T} \right]^{-}} \\ \end{array}\]

Where $\odot$ is the element-wise product. Operation $[\;]^{+}$ applied on a matrix results in a matrix with the same shape, in which all negative values are zeroed. $[\;]^{-}$ on the other hand, zeroes the positive values and takes the absolute value of negative values.

Arguments

nmf::nNMF: nNMF Struct type

Returns

χ2 : reconstruction error (frobenius norm) at last iteration

iter : exit/last iteration

Reference

Reference: Dylan Green and Stephen Bailey 2024

Moose methods

Moose.fnnls — Function

fnnls

Description

Implementation of the Fast Non-Negative Least Squares algortithm (FNNLS) adopted from the NonNegLeastSquares.jl package, (see link[https://doi.org/10.1002/(SICI)1099-128X(199709/10)11:5<393::AID-CEM483>3.0.CO;2-L]).

Problem under consideration: Ax = b, subject to x >= 0

FNNLS algorithm solves this problem in the least squares sense, || Ax - b||^2, subject to x >= 0

Arguments

AtA ::Matrix{Float32} : Gram matrix (A transpose A)

Atb ::Matrix{Float32} : Gram matrix (A transpose b)

tol ::Float32 = 1f-8 : tolerance for nonnegativity constraints

max_iter ::Int= 30*size(AtA,2) : Maximum number of iterations)

Returns

x̂ ::Vector{Float32} : x̂ that solves A*x = b in the least-squares sense, subject to `x >=0

References

Bro R, De Jong S. A fast non-negativitity-constrained least squares algorithm. Journal of Chemometrics. 11, 393–401 (1997)

Credit

Tim Holy, (NonNegLeastSquares.jl)

Moose.interpolate — Function

interpolate( basis::Basis, f::Vector{Float32}, σ::Vector{Float32}, λ::Vector{Float32}; fₓ = 0.0f0, σₓ = 1f6)

Description

This function takes an instance of the Basis Struct and three vectors: fluxes (f), standard deviations (σ) and observed wavelengths (λ), it interpolates the flux and std vectors to the rest wavelength grid assuming a redshift of 0.

Arguments

basis ::Basis : Basis Struct

f ::Vector{Float32} : Observed flux densities

σ ::Vector{Float32} : Flux associated standard deviations

λ ::Vector{Float32} : Observed wavelengths

fₓ ::Float32 = 0f0 : Extrapolation value for fluxes

σₓ ::Float32 = 1f6 : Extrapolation value for standard deviations

returns

fʳ ::Vector{Float32} : Interpolated flux densities

σʳ ::Vector{Float32} : Interpolated standard deviations

See also

interpolate!(), works similarly, the difference is that interpolate! accepts two additional arguments, fʳ and σʳ, which are the vectors of interpolated flux densities and standard deviations. Sizes of these vectors are known, so one can pass an initialization of them (e.g. with zeros) and interpolate! will fill them. This is helpful because it avoids reallocations.

Benchmark

using BenchmarkTools
@btime interpolate!(basis, f, σ, λ, fʳ , σʳ)
233.140 μs (173 allocations: 32.38 KiB)

Moose.flow — Function

flow(basis::Basis, f::Vector{Float32}, σ::Vector{Float32})

Description

This function executes a multi-threaded calculation of the chi-square curve. For each redshift (each gram matrices Hᵢ & HHᵀᵢ in basis Struct), it reconstructs an input vector f with basis vectors using a Fast Non-Negative Least Squares (FNNLS), then quantifies the reconstruction error using a χ² metric

#put definition of chi2 here grid = Γgrid()

Arguments

basis ::Basis : Basis Struct

f ::Vector{Float32} : (interpolated) flux densities vector

σ ::Vector{Float32} : (interpolated) standard deviations vector

Returns

χ² ::Vector{Float32} : χ² curve

See also

flow!()

Example

# 1. Initialize the Basis Struct
wgrid = Γgrid(λmin = 4700f0, δζ = 0.0005f0)
basis = Basis(wgrid)
# 2. Interpolate flux densities fλ and σλ to rest frame grid
fʳ , σʳ = interpolate(basis, f, σ, λ)
# 3. Run flow to get the chi2 curve
χ² = threaded_fnnls(basis, fʳ , σʳ)
#4. to get also the coeffs
χ², Ω = threaded_fnnls(basis, fʳ , σʳ, true)

Tips

To set the number of threads in julia, add this line to your linux/macos bash

export JULIA_NUM_THREADS=4

flow(wgrid::Γgrid, basis:Basis, data; output_path::String = nothing)

Description

Runs a threaded FNNLS on a set of spectra loaded into a NamedTuple apriori (data). Resulting χ² curves are saved into an .h5 file, whose path and name are specified by setting the keyword argument output_path.

Arguments

wgrid ::Γgrid : Γgrid Struct

basis ::Basis : Basis Struct

data ::Struct : Named Tuple with four fields: flux, sdev, wave, ids

output_path ::String : Path where the chi2 file will be saved

Returns

nothing

Methods

flow(data; output_path::String = nothing): wgrid and basis are instantiated within the method.

flow(fits_path::String, DataExtName::Union{String, Int}, StatExtName::Union{String, Int}; output_path::String = nothing): wgrid and basis are instantiated within the method, the data is loaded within the function using the provided fits_path and the extensions: DataExtName, StatExtName.

flow(wgrid::Γgrid, basis::Basis, fits_path::String, DataExtName::Union{String, Int}, StatExtName::Union{String, Int}; output_path::String = nothing): the data is loaded within the function using the provided fits_path and the extensions: DataExtName, StatExtName.

Example

# 1. Initialize Basis Struct
wgrid = Γgrid(;λmin, δζ)
basis = Basis(wgrid)
# 2. read data from fits files
data = leggere("../data/spectra_sample", Val(:fits))
#3. Specify output file path and run flow
chi2file = "../output/chi2_files/chi2_moose.h5"
flow(wgrid, basis, data; output_path = chi2file)

flow(src::String, ::Val{:kernel}; δζᵣ::Float32 = 0.0005f0)

Description

Reads the data from a datacube in src, and runs a threaded FNNLS on each spectrum in a sliding 3x3 mean kernel. It then writes the resulting chi-square curves into an HDF5 file. Arguments =========

src ::String : path to the HDF5 file containing the datacube

Val(:kernel) ::String : type-driven dispatch

Keyword arguments

δζᵣ ::Flot32 : test redshifts spacing for the run

Output

An HDF5 file containing the chi-square curves and other run metadata. The path to this file is generated by adding the suffix "Moose" to the source file. The HDF5 file contains an HDF5 group named "CHI2KERNEL". This group contains HDF5 datasets, each dataset corresponds to one chi-square curve with a name corresponding to its indices in the cube. For example the chi-square curve at indices (1,1) will have the name "1_1".

Returns

nothing Author(s) ========= B.Masten

flow(src::String, ::Val{:spaxel}; δζᵣ::Float32 = 0.0005f0)

Description

Reads the data from a datacube in src, and runs a threaded FNNLS on each spaxel. It then writes the resulting chi-square curves into an HDF5 file. Arguments =========

src ::String : path to the HDF5 file containing the datacube

Val(:kernel) ::String : type-driven dispatch

Keyword arguments

δζᵣ ::Flot32 : test redshifts spacing for the run

Output

An HDF5 file containing the chi-square curves and other run metadata. The path to this file is generated by adding the suffix "Moose" to the source file. The HDF5 file contains an HDF5 group named "CHI2SPAXELS". This group contains HDF5 datasets, each dataset corresponds to one chi-square curve with a name corresponding to its indices in the cube. For example the chi-square curve at indices (1,1) will have the name "1_1". Returns ======= nothing Author(s) ========= B.Masten

Utils

Moose.leggere — Function

leggere(path::String, ::Val{:fits}; DataExtName::String= "DATA", StatExtName::String ="STAT", )

Description

This function reads fluxes, standard deviations, wavelengths and ids from a group of fits file in path into a NamedTuple. Output NamedTuple has four fields (flux, sdev, awave, ids). The name of the data and stat extensions can be specified by setting the keyword arguments DataExtName and StatExtName

Arguments

path ::String : Path to the FITS files folder

Val{:fits} : Type-driven dispatch

Keyword arguments

DataExtName ::String = "DATA" : Name of the data extension

StatExtName ::String = "STAT" : Name of the stat extension

returns

data ::NamedTuple : NamedTuple with four fields (flux, sdev, awave, ids)

Warning

DataExtName and StatExtName have to be consistent across fits files

leggere(path::String, ::Val{:chi2-file})

Description

This function reads chi-square curves, best decomposition coefficients, predicted redshifts, significance scores, robustness scores, ... stored in a HDF5 file and to which the argument path points at.

Arguments

path ::String : Path to the h5 file

Val{:chi2-file} : Type-driven dispatch

returns

A NamedTuple with 6 fields (chi2, omega, zhat, dchi2, R)

leggere(path::String; DataExtName::String ="DATA", StatExtName::String = "STAT" )

Description

This function reads the flux densities, standard deviations, wavelength array and attributes and id to each spaxel in the datacube fits file in path. It outputs a NamedTuple with four fields (flux, sdev, awave, ids). The name of the data and stat extension in the fits file can be specified by setting keywords arguments DataExtName and StatExtName

Arguments

path ::String : Path to the fits files folder

DataExtName ::String = "DATA" : Name of the data extension

StatExtName ::String = "STAT" : Name of the stat extension

Returns

NamedTuple with four fields (flux, sdev, awave, ids)

Moose.fits_to_h5 — Function

fits_to_h5(fits_path::String, h5_path::Union{String,Nothing}; DataExtName::String="DATA", StatExtName::String="STAT")

Description

This is a util function that extracts the datacube content of a FITS file into an HDF5 file. This is useful as datasets in HDF5 format allow memory map.

Arguments

fits_path ::String : Path to the datacube ".fits" file

h5_path ::Union{String,Nothing} : Name/path where to save the h5 file output, if nothing a default name will be generated

Keyword Arguments

DataExtName ::String : Name of the data extension

StatExtName ::String : Name of the stat extension

Note

This function may take some time to finish, especially if the datacube has a large size (~ 13 minutes!)